FE-85 워크스페이스 사이드바 컴포넌트 분리
